requirements.txt
flask==2.3.3
pymupdf==1.23.6
gunicorn==21.2.0
python-dotenv==1.0.0
uuid==1.30
flask-cors==4.0.0
pillow==10.0.0
cryptography==41.0.1 